Changed HeartbeatCount to be ulong because the HeartbeatCount will never be < 0, it just increments.
- Also we can upcast to ulong on the KeepAlive when doing % because we
check if KeepAlive is negative via DefaultConfigurationManager.

I can only ever see this being a problem if someone were to create their
own configuration manager and then sets the KeepAlive to be negative;
very unlikely scenario.

+++ src/Microsoft.AspNet.SignalR.Core/Transports/TransportHeartBeat.cs
@@ -301,7 +301,7 @@ namespace Microsoft.AspNet.SignalR.Transports
             }
 
             // Raise keep alive if the keep alive value has passed
-            return metadata.HeartbeatCount % keepAlive == 0;
+            return metadata.HeartbeatCount % (ulong)keepAlive == 0;
         }
 
         private bool RaiseTimeout(ConnectionMetadata metadata)
@@ -371,7 +371,7 @@ namespace Microsoft.AspNet.SignalR.Transports
             public DateTime Initial { get; set; }
 
             // Number of heartbeats
-            public long HeartbeatCount { get; set; }
+            public ulong HeartbeatCount { get; set; }
 
             public void UpdateHeartbeatCount()
             {
